AP_Landing: add double casts to printf-like function

+++ libraries/AP_Landing/AP_Landing_Slope.cpp
@@ -384,7 +384,7 @@ void AP_Landing::type_slope_log(void) const
                                             type_slope_stage,
                                             flags,
                                             type_slope_flags,
-                                            slope,
-                                            initial_slope,
-                                            alt_offset);
+                                            (double)slope,
+                                            (double)initial_slope,
+                                            (double)alt_offset);
 }
